Parse into arrays of applications and services, not their configs.
