--- src/sys-config/export/WarehouseConfig.js
@@ -1,96 +1,0 @@
-// Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
-// SPDX-License-Identifier: Apache-2.0
-
-import { BaseConfig } from '@this/compote';
-
-import { ApplicationConfig } from '#x/ApplicationConfig';
-import { EndpointConfig } from '#x/EndpointConfig';
-import { HostConfig } from '#x/HostConfig';
-import { ServiceConfig } from '#x/ServiceConfig';
-
-
-/**
- * Configuration representation for the "warehouse" (the entire system).
- *
- * Accepted configuration bindings (in the constructor). All are required,
- * except as noted:
- *
- * * `{object|Array<object>} applications` -- Application configuration.
- * * `{object|Array<object>} endpoints` -- Endpoint configuration.
- * * `{object|Array<object>} hosts` -- Optional host / certificate
- *   configuration. Required only if any endpoints are configured to listen for
- *   secure connections.
- * * `{object|Array<object>} services` -- Optional system service configuration.
- *   If not present, no services are configured.
- */
-export class WarehouseConfig extends BaseConfig {
-  /**
-   * Application configuration objects.
-   *
-   * @type {Array<ApplicationConfig>}
-   */
-  #applications;
-
-  /**
-   * Host configuration objects.
-   *
-   * @type {Array<HostConfig>}
-   */
-  #hosts;
-
-  /**
-   * Endpoint configuration objects.
-   *
-   * @type {Array<EndpointConfig>}
-   */
-  #endpoints;
-
-  /**
-   * Service configuration objects.
-   *
-   * @type {Array<ServiceConfig>}
-   */
-  #services;
-
-  /**
-   * Constructs an instance.
-   *
-   * @param {object} rawConfig Configuration object. See class header for
-   *   details.
-   */
-  constructor(rawConfig) {
-    super(rawConfig);
-
-    const {
-      applications,
-      endpoints,
-      hosts = [],
-      services = []
-    } = rawConfig;
-
-    this.#applications = ApplicationConfig.parseArray(applications);
-    this.#hosts        = HostConfig.parseArray(hosts);
-    this.#endpoints    = EndpointConfig.parseArray(endpoints);
-    this.#services     = ServiceConfig.parseArray(services);
-  }
-
-  /** @returns {Array<ApplicationConfig>} Application configuration objects. */
-  get applications() {
-    return this.#applications;
-  }
-
-  /** @returns {Array<HostConfig>} Host configuration objects. */
-  get hosts() {
-    return this.#hosts;
-  }
-
-  /** @returns {Array<EndpointConfig>} Endpoint configuration objects. */
-  get endpoints() {
-    return this.#endpoints;
-  }
-
-  /** @returns {Array<ServiceConfig>} Service configuration objects. */
-  get services() {
-    return this.#services;
-  }
-}
