+++ src/sys-config/index.js
@@ -11,4 +11,3 @@ export * from '#x/RotateConfig';
 export * from '#x/SaveConfig';
 export * from '#x/ServiceConfig';
 export * from '#x/Util';
-export * from '#x/WarehouseConfig';
